Whoops forgot to push changes for months

+++ src/Graph.cpp
@@ -3,16 +3,16 @@
 #include <climits>
 
 Graph::Graph(const Graph &g) : Graph() {
-	for (uint i = 0; i < g.num_vertices; i++) add_vertex(DEFAULT_VAL);
+	for (uint32_t i = 0; i < g.num_vertices; i++) add_vertex(DEFAULT_VAL);
 
-	for (uint i = 0; i < g.num_vertices; i++) {
+	for (uint32_t i = 0; i < g.num_vertices; i++) {
 		for (auto const &[j, w] : g.edges[i]) add_directed_edge(i, j, w);
 	}
 }
 
 std::string Graph::to_string() const {
 	std::string s = "";
-	for (uint i = 0; i < num_vertices; i++) {
+	for (uint32_t i = 0; i < num_vertices; i++) {
 		s += "[" + std::to_string(vals[i]) + "] " + std::to_string(i) + " -> ";
 		for (auto const&[v, w] : edges[i]) {
 			s += "(" + std::to_string(v) + ": "  + std::to_string(w) + ") ";
@@ -22,11 +22,11 @@ std::string Graph::to_string() const {
 	return s;
 }
 
-Graph Graph::erdos_renyi_graph(uint num_vertices, float p) {
+Graph Graph::erdos_renyi_graph(uint32_t num_vertices, float p) {
 	Graph g(num_vertices);
 	thread_local std::minstd_rand r;
-	for (uint i = 0; i < num_vertices-1; i++) {
-		for (uint j = i+1; j < num_vertices; j++) {
+	for (uint32_t i = 0; i < num_vertices-1; i++) {
+		for (uint32_t j = i+1; j < num_vertices; j++) {
 			if (float(r())/float(RAND_MAX) < p)
 				g.toggle_edge(i, j);
 		}
@@ -35,14 +35,14 @@ Graph Graph::erdos_renyi_graph(uint num_vertices, float p) {
 	return g;
 }
 
-void Graph::remove_vertex(uint v) {
+void Graph::remove_vertex(uint32_t v) {
 	num_vertices--;
 	edges.erase(edges.begin() + v);
 	vals.erase(vals.begin() + v);
-	for (uint i = 0; i < num_vertices; i++) edges[i].erase(v);
+	for (uint32_t i = 0; i < num_vertices; i++) edges[i].erase(v);
 
-	for (uint i = 0; i < num_vertices; i++) {
-		std::map<uint, int> new_edges;
+	for (uint32_t i = 0; i < num_vertices; i++) {
+		std::map<uint32_t, int> new_edges;
 		for (auto const &[j, w] : edges[i]) {
 			if (j > v) new_edges.emplace(j-1, w);
 			else new_edges.emplace(j, w);
@@ -51,64 +51,64 @@ void Graph::remove_vertex(uint v) {
 	}
 }
 
-bool Graph::contains_edge(uint v1, uint v2) const {
+bool Graph::contains_edge(uint32_t v1, uint32_t v2) const {
 	return contains_directed_edge(v1, v2) && contains_directed_edge(v2, v1);
 }
 
-int Graph::edge_weight(uint v1, uint v2) const {
+int Graph::edge_weight(uint32_t v1, uint32_t v2) const {
 	return edges[v1].at(v2);
 }
 
-bool Graph::contains_directed_edge(uint v1, uint v2) const {
+bool Graph::contains_directed_edge(uint32_t v1, uint32_t v2) const {
 	return edges[v1].count(v2);
 }
 
-void Graph::set_edge_weight(uint v1, uint v2, int w) {
+void Graph::set_edge_weight(uint32_t v1, uint32_t v2, int w) {
 	edges[v1][v2] = w;
 }
 
-void Graph::add_edge(uint v1, uint v2) {
+void Graph::add_edge(uint32_t v1, uint32_t v2) {
 	add_weighted_edge(v1, v2, 1);
 }
 
-void Graph::add_directed_edge(uint v1, uint v2, int w) {
+void Graph::add_directed_edge(uint32_t v1, uint32_t v2, int w) {
 	if (!contains_edge(v1, v2)) edges[v1].emplace(v2, w);
 }
 
-void Graph::add_weighted_edge(uint v1, uint v2, int w) {
+void Graph::add_weighted_edge(uint32_t v1, uint32_t v2, int w) {
 	add_directed_edge(v1, v2, w);
 	if (v1 != v2) add_directed_edge(v2, v1, w);
 }
 
-void Graph::remove_directed_edge(uint v1, uint v2) {
+void Graph::remove_directed_edge(uint32_t v1, uint32_t v2) {
 	edges[v1].erase(v2);
 }
 
-void Graph::remove_edge(uint v1, uint v2) {
+void Graph::remove_edge(uint32_t v1, uint32_t v2) {
 	remove_directed_edge(v1, v2);
 	if (v1 != v2) remove_directed_edge(v2, v1);
 }
 
-void Graph::toggle_directed_edge(uint v1, uint v2) {
+void Graph::toggle_directed_edge(uint32_t v1, uint32_t v2) {
 	if (contains_directed_edge(v1, v2)) remove_directed_edge(v1, v2);
 	else add_directed_edge(v1, v2, 1);
 }
 
-void Graph::toggle_edge(uint v1, uint v2) {
+void Graph::toggle_edge(uint32_t v1, uint32_t v2) {
 	toggle_directed_edge(v1, v2);
 	if (v1 != v2) toggle_directed_edge(v2, v1);
 }
 
-int Graph::adjacency_matrix(uint v1, uint v2) const {
+int Graph::adjacency_matrix(uint32_t v1, uint32_t v2) const {
 	if (contains_directed_edge(v1, v2)) return edges[v1].at(v2);
 	else return 0;
 }
 
-uint Graph::degree(uint v) const {
+uint32_t Graph::degree(uint32_t v) const {
 	return edges[v].size();
 }
 
-void Graph::local_complement(uint v) {
+void Graph::local_complement(uint32_t v) {
 	for (auto const &[v1, _] : edges[v]) {
 		for (auto const &[v2, _] : edges[v]) {
 			if (v1 < v2) toggle_edge(v1, v2);
@@ -117,13 +117,13 @@ void Graph::local_complement(uint v) {
 }
 
 
-Graph Graph::partition(const std::vector<uint> &nodes) const {
-	std::set<uint> nodess;
+Graph Graph::partition(const std::vector<uint32_t> &nodes) const {
+	std::set<uint32_t> nodess;
 	std::copy(nodes.begin(), nodes.end(), std::inserter(nodess, nodess.end()));
 	Graph new_graph;
-	std::map<uint, uint> new_vertices;
+	std::map<uint32_t, uint32_t> new_vertices;
 
-	for (const uint a : nodess) {
+	for (const uint32_t a : nodess) {
 		if (!degree(a)) continue;
 
 		new_vertices.emplace(a, new_vertices.size());
@@ -144,7 +144,7 @@ Graph Graph::partition(const std::vector<uint> &nodes) const {
 
 	while (continue_deleting) {
 		continue_deleting = false;
-		for (uint i = 0; i < new_graph.num_vertices; i++) {
+		for (uint32_t i = 0; i < new_graph.num_vertices; i++) {
 			if (!new_graph.degree(i)) {
 				new_graph.remove_vertex(i);
 				continue_deleting = true;
@@ -156,15 +156,15 @@ Graph Graph::partition(const std::vector<uint> &nodes) const {
 	return new_graph;
 }
 
-std::pair<bool, std::vector<uint>> Graph::path(uint s, uint t) const {
-	std::vector<uint> stack;
+std::pair<bool, std::vector<uint32_t>> Graph::path(uint32_t s, uint32_t t) const {
+	std::vector<uint32_t> stack;
 	stack.push_back(s);
 
-	std::set<uint> visited;
-	std::map<uint, uint> parent;
+	std::set<uint32_t> visited;
+	std::map<uint32_t, uint32_t> parent;
 
 	while (!stack.empty()) {
-		uint v = *(stack.end()-1);
+		uint32_t v = *(stack.end()-1);
 		stack.pop_back();
 		if (visited.count(v)) continue;
 
@@ -175,7 +175,7 @@ std::pair<bool, std::vector<uint>> Graph::path(uint s, uint t) const {
 				if (w == t) {
 					// Done; re-use stack
 					stack.clear();
-					uint node = t;
+					uint32_t node = t;
 					while (parent.count(node)) {
 						stack.push_back(node);
 						node = parent[node];
@@ -189,16 +189,16 @@ std::pair<bool, std::vector<uint>> Graph::path(uint s, uint t) const {
 			}
 		}
 	}
-	return std::pair(false, std::vector<uint>());
+	return std::pair(false, std::vector<uint32_t>());
 }
 
-int Graph::max_flow(std::vector<uint> &sources, std::vector<uint> &sinks) const {
+int Graph::max_flow(std::vector<uint32_t> &sources, std::vector<uint32_t> &sinks) const {
 	Graph g(*this);
 
 	g.add_vertex(DEFAULT_VAL);
-	uint s = g.num_vertices - 1;
+	uint32_t s = g.num_vertices - 1;
 	g.add_vertex(DEFAULT_VAL);
-	uint t = g.num_vertices - 1;
+	uint32_t t = g.num_vertices - 1;
 
 	for (auto i : sources) g.add_directed_edge(s, i, INT_MAX);
 	for (auto i : sinks) g.add_directed_edge(i, t, INT_MAX);
@@ -208,29 +208,29 @@ int Graph::max_flow(std::vector<uint> &sources, std::vector<uint> &sinks) const 
 	return flow;
 }
 
-int Graph::max_flow(uint s, uint t) const {
+int Graph::max_flow(uint32_t s, uint32_t t) const {
 	Graph residual_graph(*this);
 
-	for (uint i = 0; i < num_vertices; i++) {
+	for (uint32_t i = 0; i < num_vertices; i++) {
 		for (auto const &[w, _] : residual_graph.edges[i]) {
 			residual_graph.add_weighted_edge(i, w, 0);
 		}
 	}
 
-	std::pair<bool, std::vector<uint>> p = residual_graph.path(s, t);
+	std::pair<bool, std::vector<uint32_t>> p = residual_graph.path(s, t);
 	bool path_exists = p.first;
-	std::vector<uint> path_nodes = p.second;
+	std::vector<uint32_t> path_nodes = p.second;
 
 	while (path_exists) {
 		int min_weight = INT_MAX;
-		for (uint j = 0; j < path_nodes.size() - 1; j++) {
+		for (uint32_t j = 0; j < path_nodes.size() - 1; j++) {
 			int weight = residual_graph.edge_weight(path_nodes[j], path_nodes[j+1]);
 			if (weight < min_weight) min_weight = weight;
 		}
 
-		for (uint j = 0; j < path_nodes.size() - 1; j++) {
-			uint u = path_nodes[j];
-			uint v = path_nodes[j+1];
+		for (uint32_t j = 0; j < path_nodes.size() - 1; j++) {
+			uint32_t u = path_nodes[j];
+			uint32_t v = path_nodes[j+1];
 			residual_graph.set_edge_weight(v, u, residual_graph.edge_weight(v, u) + min_weight);
 			residual_graph.set_edge_weight(u, v, residual_graph.edge_weight(u, v) - min_weight);
 		}
@@ -248,32 +248,32 @@ int Graph::max_flow(uint s, uint t) const {
 	return flow;
 }
 
-std::vector<uint> Graph::compute_degree_counts() const {
-	std::vector<uint> counts(num_vertices, 0);
-	for (uint i = 0; i < num_vertices; i++) counts[degree(i)]++;
+std::vector<uint32_t> Graph::compute_degree_counts() const {
+	std::vector<uint32_t> counts(num_vertices, 0);
+	for (uint32_t i = 0; i < num_vertices; i++) counts[degree(i)]++;
 	return counts;
 }
 
-std::vector<uint> Graph::compute_neighbor_degree_counts() const {
+std::vector<uint32_t> Graph::compute_neighbor_degree_counts() const {
 	// TODO thread safety?
 	std::minstd_rand rng(std::rand());
-	std::vector<uint> counts(num_vertices, 0);
-	for (uint i = 0; i < num_vertices; i++) {
+	std::vector<uint32_t> counts(num_vertices, 0);
+	for (uint32_t i = 0; i < num_vertices; i++) {
 		if (degree(i) > 0) {
-			uint v = rng() % degree(i);
+			uint32_t v = rng() % degree(i);
 			counts[degree(v)]++;
 		}
 	}
 	return counts;
 }
 
-std::set<uint> Graph::component(uint i) const {
-	std::vector<uint> stack;
+std::set<uint32_t> Graph::component(uint32_t i) const {
+	std::vector<uint32_t> stack;
 	stack.push_back(i);
-	std::set<uint> visited;
+	std::set<uint32_t> visited;
 
 	while (!stack.empty()) {
-		uint v = *(stack.end()-1);
+		uint32_t v = *(stack.end()-1);
 		stack.pop_back();
 		if (!visited.count(v)) {
 			visited.insert(v);
@@ -284,17 +284,17 @@ std::set<uint> Graph::component(uint i) const {
 	return visited;
 }
 
-float Graph::average_component_size() const {
-	std::set<uint> to_check;
-	for (uint i = 0; i < num_vertices; i++) to_check.insert(i);
-	float avg = 0.;
+double Graph::average_component_size() const {
+	std::set<uint32_t> to_check;
+	for (uint32_t i = 0; i < num_vertices; i++) to_check.insert(i);
+	double avg = 0.;
 
 	while (!to_check.empty()) {
-		uint i = *to_check.begin(); // pop
+		uint32_t i = *to_check.begin(); // pop
 		to_check.erase(to_check.begin());
 
 		auto connected_component = component(i);
-		uint component_size = connected_component.size();
+		uint32_t component_size = connected_component.size();
 		for (auto v : connected_component) {
 			if (to_check.count(v)) to_check.erase(v);
 		}
@@ -305,18 +305,18 @@ float Graph::average_component_size() const {
 	return avg/num_vertices;
 }
 
-uint Graph::max_component_size() const {
-	std::set<uint> to_check;
-	for (uint i = 0; i < num_vertices; i++) to_check.insert(i);
+uint32_t Graph::max_component_size() const {
+	std::set<uint32_t> to_check;
+	for (uint32_t i = 0; i < num_vertices; i++) to_check.insert(i);
 
-	uint max_cluster_size = 0;
+	uint32_t max_cluster_size = 0;
 
 	while (!to_check.empty()) {
-		uint i = *to_check.begin(); // pop
+		uint32_t i = *to_check.begin(); // pop
 		to_check.erase(to_check.begin());
 
 		auto connected_component = component(i);
-		uint component_size = connected_component.size();
+		uint32_t component_size = connected_component.size();
 		if (component_size > max_cluster_size) max_cluster_size = component_size;
 		for (auto v : connected_component) {
 			if (to_check.count(v)) to_check.erase(v);
@@ -326,13 +326,13 @@ uint Graph::max_component_size() const {
 	return max_cluster_size;
 }
 
-float Graph::local_clustering_coefficient(uint i) const {
-	uint ki = degree(i);
+double Graph::local_clustering_coefficient(uint32_t i) const {
+	uint32_t ki = degree(i);
 	if (ki == 0 || ki == 1) return 0.;
 
-	uint c = 0;
-	for (uint j = 0; j < num_vertices; j++) {
-		for (uint k = 0; k < num_vertices; k++) {
+	uint32_t c = 0;
+	for (uint32_t j = 0; j < num_vertices; j++) {
+		for (uint32_t k = 0; k < num_vertices; k++) {
 			c += adjacency_matrix(i, j)*adjacency_matrix(j, k)*adjacency_matrix(k, i);
 		}
 	}
@@ -340,29 +340,29 @@ float Graph::local_clustering_coefficient(uint i) const {
 	return c/(ki*(ki - 1));
 }
 
-float Graph::global_clustering_coefficient() const {
-	float c = 0.;
-	for (uint i = 0; i < num_vertices; i++) c += local_clustering_coefficient(i);
+double Graph::global_clustering_coefficient() const {
+	double c = 0.;
+	for (uint32_t i = 0; i < num_vertices; i++) c += local_clustering_coefficient(i);
 	return c/num_vertices;
 }
 
-float Graph::percolation_probability() const {
-	std::set<uint> to_check;
-	for (uint i = 0; i < num_vertices; i++) to_check.insert(i);
+double Graph::percolation_probability() const {
+	std::set<uint32_t> to_check;
+	for (uint32_t i = 0; i < num_vertices; i++) to_check.insert(i);
 
-	uint max_cluster_size = 0;
+	uint32_t max_cluster_size = 0;
 
 	while (!to_check.empty()) {
-		uint i = *to_check.begin(); // pop
+		uint32_t i = *to_check.begin(); // pop
 		to_check.erase(to_check.begin());
 
 		auto connected_component = component(i);
-		uint component_size = connected_component.size();
+		uint32_t component_size = connected_component.size();
 		if (component_size > max_cluster_size) max_cluster_size = component_size;
 		for (auto v : connected_component) {
 			if (to_check.count(v)) to_check.erase(v);
 		}
 	}
 
-	return float(max_cluster_size)/num_vertices;
+	return double(max_cluster_size)/num_vertices;
 }
